# docs/api/api-docs.yaml
openapi: 3.0.1
info:
  title: QProv API
  contact:
    name: GitHub Repository
    url: https://github.com/UST-QuAntiL/QProv
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.3.0
servers:
- url: http://localhost:5020/qprov
  description: Generated server url
paths:
...
  /provenance-documents/{provDocumentId}/entities/{provEntityId}:
    get:
      tags:
      - provenance-document
      description: Retrieve a specific PROV entity.
      operationId: getProvEntity
      parameters:
      - name: provDocumentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: provEntityId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelProvEntityDto'
        "404":
          description: Not Found. PROV entity with given ID doesn't exist.
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelProvEntityDto'
    put:
      tags:
      - provenance-document
      description: Update the entity in a specific PROV document.
      operationId: setProvEntity
      parameters:
      - name: provDocumentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: provEntityId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProvEntityDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelProvEntityDto'
        "404":
          description: Not Found. PROV document or entity with given ID doesn't exist.
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelProvEntityDto'
    delete:
      tags:
      - provenance-document
      description: Delete a PROV entity.
      operationId: deleteProvEntity
      parameters:
      - name: provDocumentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: provEntityId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "404":
          description: Not Found. PROV entity with given ID doesn't exist.
...
  /provenance-documents/{provDocumentId}/entities:
    get:
      tags:
      - provenance-document
      description: Retrieve all PROV entities of the PROV document.
      operationId: getProvEntities
      parameters:
      - name: provDocumentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelEntityModelProvEntityDto'
        "404":
          description: PROV document with the ID not available.
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelEntityModelProvEntityDto'
    post:
      tags:
      - provenance-document
      description: Create a new PROV entity in the specified PROV document.
      operationId: addProvEntityToDocument
      parameters:
      - name: provDocumentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProvQualifiedName'
        required: true
      responses:
        "201":
          description: Created
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelProvEntityDto'
        "404":
          description: Not Found. PROV document with given ID doesn't exist.
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelProvEntityDto'
...
components:
  schemas:
...
    Links:
      type: object
      additionalProperties:
        $ref: '#/components/schemas/Link'
...
    LangString:
      type: object
      properties:
        value:
          type: string
        lang:
          type: string
    Location:
      type: object
      properties:
        value:
          type: object
        type:
          $ref: '#/components/schemas/QualifiedName'
        convertedValue:
          type: object
        valueFromObject:
          type: object
          writeOnly: true
        elementName:
          $ref: '#/components/schemas/QualifiedName'
        kind:
          type: string
          enum:
          - PROV_TYPE
          - PROV_LABEL
          - PROV_ROLE
          - PROV_LOCATION
          - PROV_VALUE
          - PROV_KEY
          - OTHER
    Other:
      type: object
      properties:
        elementName:
          $ref: '#/components/schemas/QualifiedName'
        value:
          type: object
        type:
          $ref: '#/components/schemas/QualifiedName'
        convertedValue:
          type: object
        valueFromObject:
          type: object
          writeOnly: true
        kind:
          type: string
          enum:
          - PROV_TYPE
          - PROV_LABEL
          - PROV_ROLE
          - PROV_LOCATION
          - PROV_VALUE
          - PROV_KEY
          - OTHER
    ProvEntityDto:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/QualifiedName'
        label:
          type: array
          items:
            $ref: '#/components/schemas/LangString'
        location:
          type: array
          items:
            $ref: '#/components/schemas/Location'
        type:
          type: array
          items:
            $ref: '#/components/schemas/Type'
        value:
          $ref: '#/components/schemas/Value'
        others:
          type: array
          items:
            $ref: '#/components/schemas/Other'
    QualifiedName:
      type: object
      properties:
        prefix:
          type: string
        namespaceURI:
          type: string
        localPart:
          type: string
        uri:
          type: string
    Type:
      type: object
      properties:
        value:
          type: object
        type:
          $ref: '#/components/schemas/QualifiedName'
        convertedValue:
          type: object
        valueFromObject:
          type: object
          writeOnly: true
        elementName:
          $ref: '#/components/schemas/QualifiedName'
        kind:
          type: string
          enum:
          - PROV_TYPE
          - PROV_LABEL
          - PROV_ROLE
          - PROV_LOCATION
          - PROV_VALUE
          - PROV_KEY
          - OTHER
    Value:
      type: object
      properties:
        value:
          type: object
        type:
          $ref: '#/components/schemas/QualifiedName'
        convertedValue:
          type: object
        valueFromObject:
          type: object
          writeOnly: true
        elementName:
          $ref: '#/components/schemas/QualifiedName'
        kind:
          type: string
          enum:
          - PROV_TYPE
          - PROV_LABEL
          - PROV_ROLE
          - PROV_LOCATION
          - PROV_VALUE
          - PROV_KEY
          - OTHER
    EntityModelProvEntityDto:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/QualifiedName'
        label:
          type: array
          items:
            $ref: '#/components/schemas/LangString'
        location:
          type: array
          items:
            $ref: '#/components/schemas/Location'
        type:
          type: array
          items:
            $ref: '#/components/schemas/Type'
        value:
          $ref: '#/components/schemas/Value'
        others:
          type: array
          items:
            $ref: '#/components/schemas/Other'
        _links:
          $ref: '#/components/schemas/Links'
...
    ProvQualifiedName:
      type: object
      properties:
        pk:
          type: integer
          format: int64
        uri:
          type: string
        prefix:
          type: string
        namespaceURI:
          type: string
        localPart:
          type: string
        refItem:
          type: string
...
    CollectionModelEntityModelProvEntityDto:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            provEntityDtoes:
              type: array
              items:
                $ref: '#/components/schemas/EntityModelProvEntityDto'
        _links:
          $ref: '#/components/schemas/Links'
...
    Link:
      type: object
      properties:
        href:
          type: string
        hreflang:
          type: string
        title:
          type: string
        type:
          type: string
        deprecation:
          type: string
        profile:
          type: string
        name:
          type: string
        templated:
          type: boolean